Split reporting upstream client protocols
