+++ src/app/services/reporting_batch_lifecycle_service.py
@@ -6,13 +6,13 @@ from app.contracts.reporting import (
     BatchHandleResponse,
     BatchStatusResponse,
 )
+from app.services.reporting_client_protocols import ReportingBatchLifecycleClient
 from app.services.reporting_error_mapping import raise_report_batch_error
 from app.services.reporting_links import rewrite_report_batch_status_url
 from app.services.reporting_supportability import (
     RenderMetadataClient,
     attach_reporting_operator_supportability,
 )
-from app.services.upstream_client_protocols import ReportingBatchLifecycleClient
 
 
 class ReportingBatchLifecycleService:
