+++ src/app/services/reporting_client_protocols.py
@@ -1,0 +1,167 @@
+from typing import Any, Protocol
+
+
+class ReportingJobSubmissionClient(Protocol):
+    async def submit_portfolio_review_job(
+        self,
+        *,
+        payload: dict[str, Any],
+        idempotency_key: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def submit_outcome_review_report_job(
+        self,
+        *,
+        payload: dict[str, Any],
+        idempotency_key: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+
+class ReportingJobQueryClient(Protocol):
+    async def list_report_jobs(
+        self,
+        *,
+        filters: dict[str, Any],
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def get_report_job(
+        self,
+        *,
+        job_id: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def get_report_job_events(
+        self,
+        *,
+        job_id: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def get_report_job_lineage(
+        self,
+        *,
+        job_id: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def cancel_report_job(
+        self,
+        *,
+        job_id: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def get_report_snapshot(
+        self,
+        *,
+        snapshot_id: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def get_report_snapshot_lineage(
+        self,
+        *,
+        snapshot_id: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+
+class ReportingBatchSchedulerClient(Protocol):
+    async def list_report_batch_schedules(
+        self,
+        *,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def run_due_report_batch_schedules(
+        self,
+        *,
+        payload: dict[str, Any],
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+
+class ReportingBatchLifecycleClient(Protocol):
+    async def create_report_batch(
+        self,
+        *,
+        payload: dict[str, Any],
+        idempotency_key: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def get_report_batch(
+        self,
+        *,
+        batch_id: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def get_capabilities(
+        self,
+        *,
+        consumer_system: str,
+        tenant_id: str,
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+
+class ReportingBatchControlClient(Protocol):
+    async def control_report_batch(
+        self,
+        *,
+        batch_id: str,
+        action: str,
+        caller_headers: dict[str, str],
+        correlation_id: str,
+        payload: dict[str, Any] | None = None,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def get_capabilities(
+        self,
+        *,
+        consumer_system: str,
+        tenant_id: str,
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+
+class ReportingPortfolioClient(Protocol):
+    async def get_portfolio_snapshot(
+        self,
+        portfolio_id: str,
+        as_of_date: str,
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def post_portfolio_summary(
+        self,
+        portfolio_id: str,
+        payload: dict[str, Any],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
+    async def post_portfolio_review(
+        self,
+        portfolio_id: str,
+        payload: dict[str, Any],
+        correlation_id: str,
+    ) -> tuple[int, dict[str, Any]]: ...
+
